TT-5946 Mongo pkg (#459)
* changing some mongo func to public

* using common pump ProcessEnvVars

* moving mongo pump to its own pkg

* moving mongo pump to its own pkg

* renaming mongo confs

* linting uptime pump init
